Improve extendibility. #6324
