+++ easy-digital-downloads.php
@@ -130,14 +130,6 @@ final class Easy_Digital_Downloads {
 	public $components = array();
 
 	/**
-	 * EDD Back Compat handler.
-	 *
-	 * @var EDD\Back_Compat
-	 * @since 3.0.0
-	 */
-	private $back_compat;
-
-	/**
 	 * Main Easy_Digital_Downloads Instance.
 	 *
 	 * Insures that only one instance of Easy_Digital_Downloads exists in memory at any one
@@ -174,7 +166,6 @@ final class Easy_Digital_Downloads {
 			self::$instance->email_tags    = new EDD_Email_Template_Tags();
 			self::$instance->payment_stats = new EDD_Payment_Stats();
 			self::$instance->cart          = new EDD_Cart();
-			self::$instance->back_compat   = new EDD\Back_Compat();
 		}
 
 		return self::$instance;
@@ -222,7 +213,7 @@ final class Easy_Digital_Downloads {
 			case 'customers' :
 			case 'customermeta' :
 			case 'customer_meta' :
-				return new EDD\Back_Compat( 'customers' );
+				return new EDD\Compat\Customer();
 
 			default :
 				return isset( $this->{$key} )
@@ -336,6 +327,10 @@ final class Easy_Digital_Downloads {
 		require_once EDD_PLUGIN_DIR . 'includes/database/queries/class-edd-db-order-query.php';
 		require_once EDD_PLUGIN_DIR . 'includes/database/queries/class-edd-db-order-item-query.php';
 
+		// Backwards Compatibility
+		require_once EDD_PLUGIN_DIR . 'includes/compat/class-base.php';
+		require_once EDD_PLUGIN_DIR . 'includes/compat/class-customer.php';
+
 		require_once EDD_PLUGIN_DIR . 'includes/actions.php';
 		if( file_exists( EDD_PLUGIN_DIR . 'includes/deprecated-functions.php' ) ) {
 			require_once EDD_PLUGIN_DIR . 'includes/deprecated-functions.php';
@@ -354,7 +349,6 @@ final class Easy_Digital_Downloads {
 		require_once EDD_PLUGIN_DIR . 'includes/class-edd-customer.php';
 		require_once EDD_PLUGIN_DIR . 'includes/class-edd-discount.php';
 		require_once EDD_PLUGIN_DIR . 'includes/class-edd-download.php';
-		require_once EDD_PLUGIN_DIR . 'includes/class-edd-back-compat.php';
 		require_once EDD_PLUGIN_DIR . 'includes/class-edd-cache-helper.php';
 		if ( defined( 'WP_CLI' ) && WP_CLI ) {
 			require_once EDD_PLUGIN_DIR . 'includes/class-edd-cli.php';
